% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{AIC.kinetic}
\alias{AIC.kinetic}
\title{AIC for a kinetic object.}
\usage{
\method{AIC}{kinetic}(object, ...)
}
\arguments{
\item{object}{a kinetic class object.}

\item{...}{extra arguments passed to \code{\link[stats]{AIC}}.}
}
\value{
a numeric value.
}
\description{
Retrieves the value of the Akaike Information Criterion for
a kinetic class object.
}
