% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{anova.kinetic}
\alias{anova.kinetic}
\title{ANOVA for a kinetic class object.}
\usage{
\method{anova}{kinetic}(object, ...)
}
\arguments{
\item{object}{a kinetic class object.}

\item{...}{extra arguments passed to \code{\link[lme4]{anova.merMod}}.}
}
\value{
a data frame with the sum of squares, F statistic and the fraction
of explained variance.
}
\description{
Obtains an ANOVA table for a kinetic object.
}
