% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{BIC.kinetic}
\alias{BIC.kinetic}
\title{BIC for a kinetic object.}
\usage{
\method{BIC}{kinetic}(object, ...)
}
\arguments{
\item{object}{a kinetic class object.}

\item{...}{extra arguments passed to \code{\link[stats]{BIC}}.}
}
\value{
a numeric value.
}
\description{
Retrieves the value of the Bayesian Information Criterion for
a kinetic class object.
}
