% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{coef.kinetic}
\alias{coef.kinetic}
\title{Model coefficients and inference stats for a kinetic class object.}
\usage{
\method{coef}{kinetic}(object, ...)
}
\arguments{
\item{object}{a kinetic class object.}

\item{...}{extra arguments passed to the downstream functions:
\code{\link[lmerTest]{contest.lmerModLmerTest}} (gaussian family) or
\code{\link[lme4]{summary.merMod}}.}
}
\value{
a data frame with the values of model estimates, errors, CI, numbers
of complete observations and p values for the fixed-term estimates.
}
\description{
Calculates coefficient values, their errors, 95% confidence
intervals (CI) and p values, as appropriate for the modeling family.
}
\details{
In case of a CI calculation failure, the function is falling back
to computation based on the normal distribution of the estimate value.
}
