% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{deviance.kinetic}
\alias{deviance.kinetic}
\title{Deviance of a kinetic object.}
\usage{
\method{deviance}{kinetic}(object, ...)
}
\arguments{
\item{object}{a kinetic class object.}

\item{...}{extra arguments passed to \code{\link[lme4]{deviance.merMod}}.}
}
\value{
a numeric value.
}
\description{
Calculates deviance for a kinetic class object.
}
