% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{formula.kinetic}
\alias{formula.kinetic}
\title{Formula of a kinetic object.}
\usage{
\method{formula}{kinetic}(x, ...)
}
\arguments{
\item{x}{a kinetic class object.}

\item{...}{extra arguments, currently none.}
}
\value{
a formula object.
}
\description{
Extracts the model formula of a kinetic object.
}
