% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_glm_results_}
\alias{get_glm_results_}
\title{Inference statistic for a glmerMod model.}
\usage{
get_glm_results_(model, exponentiate = FALSE, ...)
}
\arguments{
\item{model}{a glmerMod mixed-effect model object.}

\item{exponentiate}{logical, should the model extimates be returned in
an exponent form?}

\item{...}{extra arguments passed to \code{\link[lme4]{summary.merMod}}.}
}
\value{
a data frame with model estimates, confidence intervals and their
significance.
}
\description{
Calculates model estimate inference stats for a GLM-like model,
more specifically for the glmerMod class.
}
\details{
In case, the default confidence interval calculation method fails,
falls back to the computation based on the normal distribution of the
estimate values.
}
