% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_lme_results_}
\alias{get_lme_results_}
\title{Retrieve inference stats or ANOVA from a merMod object.}
\usage{
get_lme_results_(model, L = NULL, joint = FALSE, ...)
}
\arguments{
\item{model}{a merMod mixed-effect model object.}

\item{L}{a contrast vector or a matrix. If NULL, inference stats for all
model estimates are returned.}

\item{joint}{logical, make an F-test of potentially several contrast vectors?
If FALSE single DF t-tests are applied to each vector or each row of
contrasts matrices.}

\item{...}{additional arguments passed to
\code{\link[lmerTest]{contest.lmerModLmerTest}}.}
}
\value{
a data frame with model estimates, confidence intervals and their
significance or, if joint = TRUE, an ANOVA table.
}
\description{
Obtains inference statistics or ANOVA table for a mixed-effect
model (class merMod).
}
\details{
a wrapper around
\code{\link[lmerTest]{contest.lmerModLmerTest}} function.
}
