% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_qc_plots}
\alias{get_qc_plots}
\title{Plots of model residuals.}
\usage{
get_qc_plots(
  kinetic_object,
  cust_theme = ggplot2::theme_classic(),
  resid.type = "response",
  point_wjitter = 0.01,
  point_hjitter = 0.01,
  point_alpha = 0.75,
  ...
)
}
\arguments{
\item{cust_theme}{a ggplot theme.}

\item{resid.type}{type of the residuals, passed to
\code{\link{residuals.kinetic}}.}

\item{point_wjitter}{horizontal point jittering.}

\item{point_hjitter}{vertical point jittering.}

\item{point_alpha}{plot point alpha.}

\item{...}{extra arguments passed to \code{\link{residuals.kinetic}}.}

\item{kinet_object}{a kinetic object.}
}
\description{
Generates a series of residual plots: residuals vs fitted,
standardized residuals vs fitted, squared residuals vs fitted and residuals
quantile-quantile plot.
}
