% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_lrt.R
\name{get_stats}
\alias{get_stats}
\title{Fit stats for a kinetic class object.}
\usage{
get_stats(kinetic_object, ...)
}
\arguments{
\item{kinetic_object}{a kinetic class object.}

\item{...}{additional arguments passed to \code{\link{residuals.kinetic}}.}
}
\value{
a data frame with the n number of complete observations, AIC, BIC,
R squared, deviance, mae (mean absolute error), mse (mean squared error) and
rmse (root mean squared error).
}
\description{
Calculates fit errors, AIC, BIC and
raw pseudo-R squares (1 - MSE/Var(y)) for a kinetic model.
}
