% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{identity_matrix_}
\alias{identity_matrix_}
\title{Generate an identity matrix.}
\usage{
identity_matrix_(k)
}
\arguments{
\item{k}{matrix dimension.}
}
\value{
a matrix.
}
\description{
Creates an identity matrix of the given dimension.
}
