% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_lrt.R
\name{inference_kinetic_list}
\alias{inference_kinetic_list}
\title{Inference stats for the kinetic object list.}
\usage{
inference_kinetic_list(
  kinetic_objects,
  .parallel = FALSE,
  .errors = FALSE,
  ...
)
}
\arguments{
\item{kinetic_objects}{a list of kinetic class objects.}

\item{.parallel}{logical, should the model construction be done in parallel?}

\item{.errors}{logical, should modeling failures be returned?}

\item{...}{extra arguments passed to \code{\link{coef.kinetic}}.}
}
\value{
a list of data frames with the inference stats.
}
\description{
Calculates inference stats for a list of kinetic objects,
optionally in parallel. See: \code{\link{coef.kinetic}} for details.
}
