% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{is_kinetic}
\alias{is_kinetic}
\title{Check for the kinetic_object class.}
\usage{
is_kinetic(x)
}
\arguments{
\item{x}{an object.}
}
\value{
a logical value.
}
\description{
Checks if the object is an instance of the 'kinetic' class.
}
