% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{kinetic}
\alias{kinetic}
\title{Make a kinetic class object.}
\usage{
kinetic(model, response, ID, time, family, order)
}
\arguments{
\item{model}{a merMod model.}

\item{response}{the name of the dependent variable.}

\item{ID}{the name of the mixed-effect variable, i.e. specifying the
data pairing/blocking.}

\item{time}{the name of the time variable.}

\item{family}{the model family name (GLM).}

\item{order}{model order.}
}
\value{
an instance of the 'kinetic' class.
}
\description{
Generates an instance of the 'kinetic' class given a merMod
model and it's base features.
}
