% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_lrt.R
\name{lrt}
\alias{lrt}
\title{Likelihood ratio test for a kinetic class object.}
\usage{
lrt(kinetic_object, ...)
}
\arguments{
\item{kinetic_object}{a kinetic class object.}

\item{...}{additional arguments passed to \code{\link[stats]{anova}}.}
}
\value{
a data frame with the likelihood, AIC, BIC values and the test
statistics.
}
\description{
Performs a likelihood ratio test for a kinetic class object.
Model terms are compared in the 0 - n order; additionally, the full model is
compared with the NULL model (order = 'global').
}
\details{
The effect size statistic lambda is calculated with the formula
2 * (logLik(model) - logLik(NULL model))
}
