% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_lrt.R
\name{lrt_list}
\alias{lrt_list}
\title{Likelihood ration test for a list of kinetic objects.}
\usage{
lrt_list(kinetic_objects, .parallel = FALSE, .errors = FALSE, ...)
}
\arguments{
\item{kinetic_objects}{a list with kinetic class objects.}

\item{.parallel}{logical, should the model construction be done in parallel?}

\item{.errors}{logical, should modeling failures be returned?}

\item{...}{extra arguments passed to \code{\link[stats]{anova}}.}
}
\value{
a list of data frames with the likelihood, AIC, BIC values and
the test statistics.
}
\description{
Performs likelihood ratio tests for a list of kinetic class
objects, optionally in parallel.
}
\details{
The effect size statistic lambda is calculated with the formula
2 * (logLik(model) - logLik(NULL model))
}
