% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{model.frame.kinetic}
\alias{model.frame.kinetic}
\title{Model frame of the kinetic object.}
\usage{
\method{model.frame}{kinetic}(formula, ...)
}
\arguments{
\item{formula}{a kinetic class object.}

\item{...}{extra arguments, currently none.}
}
\value{
a data frame.
}
\description{
Extracts the kinetic model data.
}
