% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{model_kinetic_lst}
\alias{model_kinetic_lst}
\title{Generate a list of kinetic objects.}
\usage{
model_kinetic_lst(
  data,
  responses = NULL,
  time = "time",
  ID = "ID",
  family = "gaussian",
  order = 2,
  .parallel = FALSE,
  .errors = FALSE,
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{time}{name of the time variable.}

\item{ID}{name of the pairing/grouping or block variable.}

\item{family}{modeling family.}

\item{.parallel}{logical, should the model construction be done in parallel?}

\item{.errors}{logical, should modeling failures be returned?}

\item{...}{extra arguments passed to \code{\link[lme4]{lmer}} or
\code{\link[lme4]{glmer}}.}

\item{response}{a vector with the response variable names.}
}
\value{
a list with kinetic class objects, one for each response variable.
}
\description{
Generates a list of kinetic objects for a vector of response
variable names. A wrapper around \code{\link{model_kinetic}} enabling for
model construction in parallel.
}
