% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{nobs.kinetic}
\alias{nobs.kinetic}
\title{Get observation numbers for a kinetic object.}
\usage{
\method{nobs}{kinetic}(object, ...)
}
\arguments{
\item{object}{a kinetic class object.}

\item{...}{extra arguments, currently none.}
}
\description{
Extracts the number of complete observations..
}
