% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_kin_fct_}
\alias{plot_kin_fct_}
\title{Plot frequency of a binary feature.}
\usage{
plot_kin_fct_(
  kinetic_object,
  point_color = "steelblue",
  outcome_color = "steelblue",
  fitted_color = "coral3",
  fitted_line = "dashed",
  cust_theme = ggplot2::theme_classic(),
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  x_lab = kinetic_object$time,
  y_lab = "\% cohort",
  labels = TRUE,
  repel_labels = FALSE,
  signif_digits = 2,
  label_size = 2.3,
  fitted = FALSE
)
}
\arguments{
\item{kinetic_object}{a kinetic class object.}

\item{point_color}{color of the points and text labels.}

\item{outcome_color}{color of the median and IQR representations.}

\item{fitted_color}{color of the model fit line.}

\item{fitted_line}{style of the model fit line.}

\item{cust_theme}{a ggplot theme.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle.}

\item{plot_tag}{plot tag.}

\item{x_lab}{x axis title.}

\item{y_lab}{y axis title.}

\item{labels}{logical, display percent labels for the time points?}

\item{repel_labels}{logical, place the labels to reduce their overlap?}

\item{signif_digits}{significant digits for the percent value rounding.}

\item{label_size}{size of the percent labels.}

\item{fitted}{logical, display the model fit?}
}
\value{
a ggplot object.
}
\description{
Plots the frequency of a binary feature together with the
model predictions.
}
