% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_kin_numeric_}
\alias{plot_kin_numeric_}
\title{Generate a trajectory plot for a kinetic object.}
\usage{
plot_kin_numeric_(
  kinetic_object,
  point_color = "gray60",
  point_alpha = 0.5,
  outcome_color = "steelblue",
  fitted_color = "coral3",
  fitted_line = "dashed",
  cust_theme = ggplot2::theme_classic(),
  jitter_w_perc = 2,
  jitter_h_perc = 1,
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  x_lab = kinetic_object$time,
  y_lab = kinetic_object$response,
  IQR = TRUE,
  fitted = TRUE,
  IQR_fitted = FALSE
)
}
\arguments{
\item{kinetic_object}{a kinetic class object.}

\item{point_color}{color of the points (observations).}

\item{point_alpha}{point alpha.}

\item{outcome_color}{color of the median and IQR representations.}

\item{fitted_color}{color of the model fit line.}

\item{fitted_line}{style of the model fit line.}

\item{cust_theme}{a ggplot theme.}

\item{jitter_w_perc}{horizontal jittering of the point, in percents.}

\item{jitter_h_perc}{vertical jittering of the points, in percents.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle.}

\item{plot_tag}{plot tag.}

\item{x_lab}{x axis title.}

\item{y_lab}{y axis title.}

\item{IQR}{logical, display the IQR of the outcome?}

\item{fitted}{logical, display the model fit?}

\item{IQR_fitted}{logical, display the model fit IQR?}
}
\value{
a ggplot object.
}
\description{
Plots a numeric kinetic object. Points represent single
observations, medians and IQR (interquartile ranges) at particular time
points are displayed as lines and colored regions.
}
