% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{print.kinetic}
\alias{print.kinetic}
\title{Print a kinetic object.}
\usage{
\method{print}{kinetic}(x, ...)
}
\arguments{
\item{x}{a kinetic class object.}

\item{...}{extra arguments, currently none.}
}
\value{
none, called for it's side effects.
}
\description{
Prints a kinetic object.
}
