% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{residuals.kinetic}
\alias{residuals.kinetic}
\title{Residuals of a kinetic object.}
\usage{
\method{residuals}{kinetic}(object, ...)
}
\arguments{
\item{object}{a kinetic class object.}

\item{...}{extra arguments passed to
\code{\link[lme4]{residuals.merMod}}.}
}
\value{
a tibble with the actual and predicted values and the residuals.
}
\description{
Extracts model residuals from a kinetic object.
}
\details{
Potential missfits are identified with the +/- 2*SD rule.
}
