% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{summary.kinetic}
\alias{summary.kinetic}
\title{Model summary for a kinetic class object.}
\usage{
\method{summary}{kinetic}(
  object,
  type = c("inference", "fit", "assumptions"),
  resid.type = "response",
  ...
)
}
\arguments{
\item{object}{a kinetic class object.}

\item{type}{type of the statistic returned.}

\item{resid.type}{type of the residuals, passed to
\code{\link{residuals.kinetic}}.}

\item{...}{extra arguments passed to \code{\link{coef.kinetic}}
or \code{\link{residuals.kinetic}}.}
}
\description{
Retrieves inference or fit statistics or normality test results
for the model residuals from a kinetic class object.
}
